Specify limits of TSC membership
This change specifies a limit of 5 members of TSC at any one time.

This was originally included in #48 but was removed to prevent blocking
the rest of the PR which was more focused on the RFC process.

I want to champion this change because we are looking to expand the TSC
outside of the 2 current members, but have learned from other
organizations (Node.js and ESLint in particular) that having a small TSC
is helpful to keep things moving quickly.